More work
More work on sessions and login. Added a users table to the site db for
better handling of user info to not always fetch from robust. Moved
bootstrap css to a "template" folder for more customization. Added
fields in the user and settings table to support customization. See im
still working on this even though im more back in SL again.