Major update
Login now working but had to rewrite it so the password encryption will
be different then opensim encryption, this was the only way, also its a
good security feature i think. Introduction to the market place clone
idea. Touch ups and fixes all around. Install of webmaps, font-awesome
and many more with more to come.