API redo
Just redid the API system for it to be more RESTful friendly. Still
needs fixing and testing but examples for both PHP and LSL are included.
Register is now implimented but there are problems that i am unable to
find and default avatar system not writen yet. Its a crude way to insert
new residents but should work.